+++ inc/weapons.h
@@ -1,6 +1,6 @@
 /*
  * weapons.h
- * Copyright (C) 2009-2018 Joachim de Groot <jdegroot@web.de>
+ * Copyright (C) 2009-2025 Joachim de Groot <jdegroot@web.de>
  *
  * NLarn is free software: you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -77,6 +77,7 @@ typedef enum _weapon_t
     WT_BATTLEAXE,
     WT_CROSSBOW,
     WT_LONGSWORD,
+    WT_ELONGSWORD,
     WT_2SWORD,
     WT_SWORDSLASHING,
     WT_LANCEOFDEATH,